third_party/blink/web_tests/external/wpt/svg/struct/systemLanguage-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] Leading and trailing whitespace on the value is stripped
  assert_equals: token count expected 2 but got 0
[FAIL] Tabs and newlines as leading/trailing whitespace are stripped
  assert_equals: token count expected 2 but got 0
[FAIL] Double comma produces an empty token between the two commas
  assert_equals: token count expected 3 but got 1
[FAIL] Empty string results in a single empty token
  assert_equals: token count expected 1 but got 0
[FAIL] A single comma separates two empty tokens
  assert_equals: token count expected 2 but got 0
Harness: the test ran to completion.
